Added roborally as a field in GameGraphics
