+++ src/main/java/no/uib/inf112/core/RoboRally.java
@@ -19,7 +19,7 @@ import java.io.File;
 import java.util.concurrent.Executors;
 import java.util.concurrent.ScheduledExecutorService;
 
-public class RoboRally extends Game {
+public class RoboRally  {
 
     public static final String MAP_FOLDER = "maps";
     public static final int PHASES_PER_ROUND = 5;
@@ -27,41 +27,16 @@ public class RoboRally extends Game {
     //see https://github.com/inf112-v19/Blodtorstige-robotet/wiki/Run-with-IntelliJ
     public static final String FALLBACK_MAP_FILE_PATH = MAP_FOLDER + File.separatorChar + "test.tmx";
 
-    private SpriteBatch batch;
-    private BitmapFont font;
 
-    private static TiledMapHandler map;
 
-    private static PlayerHandler playerHandler;
+    protected static TiledMapHandler map;
+
+    protected static PlayerHandler playerHandler;
     public static MapInteractOnUser mapInteractOnUser;
 
-    private static InputMultiplexer inputMultiplexer;
-    private static UIHandler uiHandler;
-    private static ControlPanelEventHandler cpEventHandler;
-    public static ScheduledExecutorService executorService;
 
-    @Override
-    public void create() {
 
-        executorService = Executors.newSingleThreadScheduledExecutor();
 
-        batch = new SpriteBatch();
-        font = new BitmapFont();
-
-        inputMultiplexer = new InputMultiplexer();
-        Gdx.input.setInputProcessor(inputMultiplexer);
-
-        cpEventHandler = new ControlPanelEventHandler();
-        map = new TiledMapHandler(FALLBACK_MAP_FILE_PATH);
-
-        playerHandler = new PlayerHandler(3);
-        playerHandler.generatePlayers(false);
-        mapInteractOnUser = new MapInteractOnUser();
-        uiHandler = new UIHandler();
-        new InputHandler(); //this must be after UIHandler to allow dragging of cards
-
-        playerHandler.doTurn();
-    }
 
     public static void round() {
         for (int i = 0; i < PHASES_PER_ROUND; i++) {
@@ -79,37 +54,8 @@ public class RoboRally extends Game {
         //User plans next round
     }
 
-    @Override
-    public void render() {
-        Gdx.gl.glClear(GL20.GL_COLOR_BUFFER_BIT |
-                (Gdx.graphics.getBufferFormat().coverageSampling ? GL20.GL_COVERAGE_BUFFER_BIT_NV : 0));
 
-        super.render();
 
-        batch.begin();
-
-        map.update(Gdx.graphics.getDeltaTime());
-        map.render(batch);
-
-        uiHandler.update();
-
-        batch.end();
-    }
-
-    @Override
-    public void dispose() {
-        super.dispose();
-        batch.dispose();
-        font.dispose();
-        uiHandler.dispose();
-    }
-
-    @Override
-    public void resize(int width, int height) {
-        super.resize(width, height);
-        map.resize();
-        uiHandler.resize();
-    }
 
     /**
      * @return The current map in play
@@ -119,23 +65,10 @@ public class RoboRally extends Game {
         return map;
     }
 
-    @NotNull
-    public static InputMultiplexer getInputMultiplexer() {
-        return inputMultiplexer;
-    }
 
-    @NotNull
-    public static ControlPanelEventHandler getCPEventHandler() {
-        return cpEventHandler;
-    }
 
     @NotNull
     public static PlayerHandler getPlayerHandler() {
         return playerHandler;
-    }
-
-    @NotNull
-    public static UIHandler getUiHandler() {
-        return uiHandler;
     }
 }
