+++ src/main/java/no/uib/inf112/desktop/Main.java
@@ -2,6 +2,7 @@ package no.uib.inf112.desktop;
 
 import com.badlogic.gdx.backends.lwjgl.LwjglApplication;
 import com.badlogic.gdx.backends.lwjgl.LwjglApplicationConfiguration;
+import no.uib.inf112.core.GameGraphics;
 import no.uib.inf112.core.RoboRally;
 
 
@@ -14,6 +15,6 @@ public class Main {
         cfg.height = 720;
         cfg.samples = 16; //max out the samples as this isn't a very heavy game.
 
-        new LwjglApplication(new RoboRally(), cfg);
+        new LwjglApplication(new GameGraphics(), cfg);
     }
 }
